Move slang-com-ptr.h to be publically available.
