+++ tools/render-test/resource-d3d12.h
@@ -1,4 +1,4 @@
-// resource-d3d12.h
+﻿// resource-d3d12.h
 #pragma once
 
 #define WIN32_LEAN_AND_MEAN
@@ -10,7 +10,7 @@
 #include <dxgi1_4.h>
 #include <d3d12.h>
 
-#include "../../source/core/slang-com-ptr.h"
+#include "../../slang-com-ptr.h"
 #include "d3d-util.h"
 
 namespace renderer_test {
